server/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
python-dotenv==0.21.0
